requirements.txt
asgiref==3.6.0
audioread==3.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
decorator==5.1.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
emosic-essentials @ git+https://github.com/Anupam96786/emosic-essentials.git@7304339af5cb18d6714fd31ab0dc64895fbba204
idna==3.4
joblib==1.2.0
lazy_loader==0.2
librosa==0.10.0
llvmlite==0.40.0
msgpack==1.0.5
numba==0.57.0
numpy==1.24.3
packaging==23.1
platformdirs==3.5.1
pooch==1.7.0
pycparser==2.21
pytz==2023.3
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.5
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==2.0.2
